LPS-95716 Move CleanUpPermissionsUtil and CleanUpPortletPreferencesUtil to module
